pkg/trace/api: add v0.5 endpoint with more compact data format
This change adds a new endpoint (/v0.5/traces) which supports receiving
spans as an array formed of its 12 properties (avoiding the string
duplication of each field name).

Spans are encoded as a 12-element array containing only the values for
each field, in this exact order:

  [ Service, Name, Resource, TraceID, SpanID, ParentID, Start, Duration, Error, Meta, Metrics, Type ]

Empty fields will be nil.